checking accessablity for all the methods of role_controller class...

+++ src/main/java/com/example/lib_management/controller/Roleservice_controller.java
@@ -1,6 +1,7 @@
 package com.example.lib_management.controller;
 
 import com.example.lib_management.AAA.Entity.Aaaccesslist;
+import com.example.lib_management.AAA.ServiceInterface.LoadPageWebService;
 import com.example.lib_management.AAA.ServiceInterface.RoleWebService;
 import com.example.lib_management.AAA.commons.exceptions.mException;
 
@@ -20,11 +21,14 @@ import java.util.List;
 @Controller
 @RequestMapping("rolecontroller")
 public class Roleservice_controller {
+
    private RoleWebService roleService;
+   private LoadPageWebService loadPageWebService;
 
     @Autowired
-    public Roleservice_controller(RoleWebService roleService){
+    public Roleservice_controller(RoleWebService roleService , LoadPageWebService loadPageWebService){
         this.roleService=roleService;
+        this.loadPageWebService=loadPageWebService;
     }
 
 
@@ -32,24 +36,19 @@ public class Roleservice_controller {
     @RequestMapping(value = "addrole",method = RequestMethod.GET)
     public String add_role_page(@ModelAttribute ("role_key") RoleinfoDTO roleinfoDTO , HttpServletRequest request) throws mException {
 
-        // checking if this page is accessable or not
-        HttpSession httpSession= request.getSession(false);
-        if(httpSession==null)
-            return "redirect:/logincontroller/login";
-
-        List<Aaaccesslist> aaaccesslists = (List<Aaaccesslist>) httpSession.getAttribute("AccessList");
-
-        for (Aaaccesslist aaaccesslist:aaaccesslists)
-        {
-            if (aaaccesslist.getAapage().getUrl().equals("http://localhost:8080/rolecontroller/addrole"))
-                return "add_role";
-        }
-
+        if (loadPageWebService.check_access_to_page("http://localhost/rolecontroller/addrole",request))
+            return "add_role";
         return "redirect:/logincontroller/login";
     }
 
     @RequestMapping(value = "submit_addrole_form")
-    public String add_role_submit(RoleinfoDTO roleinfoDTO) throws mException {
+    public String add_role_submit(RoleinfoDTO roleinfoDTO , HttpServletRequest request) throws mException
+    {
+        ////////////// chcking accessability
+        if (!(loadPageWebService.check_access_to_page("http://localhost/rolecontroller/addrole",request)))
+            return "redirect:/logincontroller/login";
+
+
       String message=  roleService.add_role(roleinfoDTO);
       System.out.println(message);
 
@@ -58,18 +57,26 @@ public class Roleservice_controller {
 //////////////////////////////////////////////////////////////////////////// Show All Roles
 
     @RequestMapping(value = "viewallroles", method = RequestMethod.GET)
-    public String roles_info_page(Model model)
+    public String roles_info_page(Model model , HttpServletRequest request)
     {
-        System.out.println("entering viewallroles...");
+        /// chcking accessability
+        if (!(loadPageWebService.check_access_to_page("http://localhost/rolecontroller/viewallroles",request)))
+            return "redirect:/logincontroller/login";
+
+
+
         model.addAttribute("roles_list_key", roleService.return_all_roles());
 
         return "view_all_roles";
     }
 //////////////////////////////////////////////////////////////////////////// Delete Role
 
-    @RequestMapping(value = "deleterole/{code}",method = RequestMethod.GET)
-    public String delete_role(@PathVariable int code){
-        System.out.println("inside delete method with code:"+code);
+    @RequestMapping(value = "deleterole/{code}",method = RequestMethod.GET )
+    public String delete_role(@PathVariable int code ,HttpServletRequest request) {
+        /// chcking accessability
+        if (!(loadPageWebService.check_access_to_page("http://localhost/rolecontroller/viewallroles",request)))
+            return "redirect:/logincontroller/login";
+
         roleService.Remove(code);
 
 
@@ -78,7 +85,14 @@ public class Roleservice_controller {
 //////////////////////////////////////////////////////////////////////////// Edit Role
 
     @RequestMapping(value = "editrole/{code}" ,method = RequestMethod.GET)
-    public String show_edit_role_form(@PathVariable int code,Model model){
+    public String show_edit_role_form(@PathVariable int code,Model model , HttpServletRequest request){
+
+        /// chcking accessability
+        if (!(loadPageWebService.check_access_to_page("http://localhost/rolecontroller/editrole",request)))
+            return "redirect:/logincontroller/login";
+
+
+
         RoleinfoDTO roleinfoDTO=roleService.return_roel_by_id(code);
         model.addAttribute("role_key",roleinfoDTO);
 
@@ -87,7 +101,13 @@ public class Roleservice_controller {
 
 
     @RequestMapping(value ="submit_edit_role/submit_edit_page",method = RequestMethod.POST)
-    public String submit_edit_role_form(RoleinfoDTO roleinfoDTO){
+    public String submit_edit_role_form(RoleinfoDTO roleinfoDTO,HttpServletRequest request){
+
+        /// chcking accessability
+        if (!(loadPageWebService.check_access_to_page("http://localhost/rolecontroller/editrole",request)))
+            return "redirect:/logincontroller/login";
+
+
         String message=  roleService.update_role(roleinfoDTO);
         System.out.println(message);
